% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-common.R
\name{audio_info}
\alias{audio_info}
\title{Audio Information}
\usage{
audio_info(filepath)
}
\arguments{
\item{filepath}{(str) path to the mp3/wav file.}
}
\value{
data.frame.
}
\description{
Audio Information
}
