% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-utils.R
\name{backend_utils_list_audio_backends}
\alias{backend_utils_list_audio_backends}
\title{List Available Audio Backends}
\usage{
backend_utils_list_audio_backends()
}
\value{
character vector with the list of available backends.
}
\description{
List Available Audio Backends
}
