% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-cmuarctic.R
\name{cmuarctic_dataset}
\alias{cmuarctic_dataset}
\title{CMU Arctic Dataset}
\usage{
cmuarctic_dataset(
  root,
  url = "aew",
  folder_in_archive = "ARCTIC",
  download = FALSE
)
}
\arguments{
\item{root}{(str): Path to the directory where the dataset is found or downloaded.}

\item{url}{(str, optional): The URL to download the dataset from or the type of the dataset to dowload.
(default: \code{"aew"})
Allowed type values are \code{"aew"}, \code{"ahw"}, \code{"aup"}, \code{"awb"}, \code{"axb"}, \code{"bdl"},
\code{"clb"}, \code{"eey"}, \code{"fem"}, \code{"gka"}, \code{"jmk"}, \code{"ksp"}, \code{"ljm"}, \code{"lnh"},
\code{"rms"}, \code{"rxr"}, \code{"slp"} or \code{"slt"}.}

\item{folder_in_archive}{(str, optional): The top-level directory of the dataset.  (default: \code{"ARCTIC"})}

\item{download}{(bool, optional): Whether to download the dataset if it is not found at root path.  (default: \code{FALSE}).}
}
\value{
a torch::dataset()
}
\description{
Create a Dataset for CMU_ARCTIC.
}
