% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{complex_norm}
\alias{complex_norm}
\title{Complex Norm}
\usage{
complex_norm(complex_tensor, power = 1)
}
\arguments{
\item{complex_tensor}{(tensor): Tensor shape of \verb{(..., complex=2)}}

\item{power}{(numeric): Power of the norm. (Default: \code{1.0}).}
}
\value{
tensor: Power of the normed input tensor. Shape of \verb{(..., )}
}
\description{
Compute the norm of complex tensor input.
}
