% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models-convtasnet.R
\name{.align_num_frames_with_strides}
\alias{.align_num_frames_with_strides}
\title{Align frames with strides}
\usage{
.align_num_frames_with_strides(input, enc_kernel_size, enc_stride)
}
\arguments{
\item{input}{(Tensor): 3D Tensor with shape (batch_size, channels==1, frames)}

\item{enc_kernel_size}{(int): The convolution kernel size of the encoder/decoder, L.}

\item{enc_num_feats}{(int): The feature dimensions passed to mask generator, N.}
}
\value{
list(padded_tensor, number_of_paddings_performed)
}
\description{
Pad input Tensor so that the end of the input tensor corresponds with.
}
\section{Assumption}{


The resulting Tensor will be padded with the size of stride  (== kernel_width // 2)
on the both ends in Conv1D\preformatted{   |<--- k_1 --->|
   |      |            |<-- k_n-1 -->|
   |      |                  |  |<--- k_n --->|
   |      |                  |         |      |
   |      |                  |         |      |
   |      v                  v         v      |
   |<---->|<--- input signal --->|<--->|<---->|
    stride                         PAD  stride
}
\enumerate{
\item (If kernel size is odd) the center of the last convolution kernel.
\item (If kernel size is even) the end of the first half of the last convolution kernel.
}
}

