% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{functional__combine_max}
\alias{functional__combine_max}
\title{Combine Max (functional)}
\usage{
functional__combine_max(a, b, thresh = 0.99)
}
\arguments{
\item{a}{(list(tensor, tensor))}

\item{b}{(list(tensor, tensor))}

\item{thresh}{(float) Default: 0.99}
}
\value{
\code{list(tensor, tensor)}: a list with values tensor and indices tensor.
}
\description{
Take value from first if bigger than a multiplicative factor of the second, elementwise.
}
