% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{functional__compute_nccf}
\alias{functional__compute_nccf}
\title{Normalized Cross-Correlation Function (functional)}
\usage{
functional__compute_nccf(waveform, sample_rate, frame_time, freq_low)
}
\arguments{
\item{waveform}{(Tensor): Tensor of audio of dimension (..., time)}

\item{sample_rate}{(int): sampling rate of the waveform, e.g. 44100 (Hz)}

\item{frame_time}{(float)}

\item{freq_low}{(float)}
}
\value{
\code{tensor} of nccf``
}
\description{
Compute Normalized Cross-Correlation Function  (NCCF).
}
