% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{functional__generate_wave_table}
\alias{functional__generate_wave_table}
\title{Wave Table Generator (functional)}
\usage{
functional__generate_wave_table(
  wave_type,
  data_type,
  table_size,
  min,
  max,
  phase,
  device
)
}
\arguments{
\item{wave_type}{(str): 'SINE' or 'TRIANGULAR'}

\item{data_type}{(str): desired data_type ( \code{INT} or \code{FLOAT} )}

\item{table_size}{(int): desired table size}

\item{min}{(float): desired min value}

\item{max}{(float): desired max value}

\item{phase}{(float): desired phase}

\item{device}{(torch_device): Torch device on which table must be generated}
}
\value{
\code{tensor}: A 1D tensor with wave table values
}
\description{
A helper function for phaser. Generates a table with given parameters
}
