% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{functional_allpass_biquad}
\alias{functional_allpass_biquad}
\title{All-pass Biquad Filter}
\usage{
functional_allpass_biquad(waveform, sample_rate, central_freq, Q = 0.707)
}
\arguments{
\item{(torch.Tensor):}{audio waveform of dimension of \verb{(..., time)}}

\item{(int):}{sampling rate of the waveform, e.g. 44100 (Hz)}

\item{(float):}{central frequency (in Hz)}

\item{(float, }{optional): \url{https://en.wikipedia.org/wiki/Q_factor} (Default: \code{0.707})}
}
\value{
Tensor: Waveform of dimension of \verb{(..., time)}
}
\description{
Design two-pole all-pass filter. Similar to SoX implementation.
}
\references{
\itemize{
\item \url{http://sox.sourceforge.net/sox.html}
\item \url{https://www.w3.org/2011/audio/audio-eq-cookbook.html#APF}
}
}
