% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{functional_angle}
\alias{functional_angle}
\title{Angle (functional)}
\usage{
functional_angle(complex_tensor)
}
\arguments{
\item{complex_tensor}{(Tensor): Tensor shape of \verb{(..., complex=2)}}
}
\value{
\code{tensor}: Angle of a complex tensor. Shape of \verb{(..., )}
}
\description{
Compute the angle of complex tensor input.
}
