% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{functional_combine_max}
\alias{functional_combine_max}
\title{Combine Max}
\usage{
functional_combine_max(a, b, thresh = 0.99)
}
\arguments{
\item{a}{(Tuple\link{Tensor, Tensor})}

\item{b}{(Tuple\link{Tensor, Tensor})}

\item{thresh}{(float) Default: 0.99}
}
\description{
Take value from first if bigger than a multiplicative factor of the second, elementwise.
}
