% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{functional_compute_nccf}
\alias{functional_compute_nccf}
\title{Normalized Cross-Correlation Function (functional)}
\usage{
functional_compute_nccf(waveform, sample_rate, frame_time, freq_low)
}
\arguments{
\item{waveform}{(Tensor): Tensor of audio of dimension (..., time)}

\item{sample_rate}{(int): Sample rate of the audio waveform}

\item{frame_time}{(float)}

\item{freq_low}{(float)}
}
\description{
Compute Normalized Cross-Correlation Function  (NCCF).
}
