% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{functional_contrast}
\alias{functional_contrast}
\title{Contrast Effect}
\usage{
functional_contrast(waveform, enhancement_amount = 75)
}
\arguments{
\item{waveform}{(Tensor): audio waveform of dimension of \verb{(..., time)}}

\item{enhancement_amount}{(float): controls the amount of the enhancement
Allowed range of values for enhancement_amount : 0-100
Note that enhancement_amount = 0 still gives a significant contrast enhancement}
}
\value{
\code{tensor}: Waveform of dimension of \verb{(..., time)}
}
\description{
Apply contrast effect.  Similar to SoX implementation.
Comparable with compression, this effect modifies an audio signal to
make it sound louder
}
\references{
\itemize{
\item \url{http://sox.sourceforge.net/sox.html}
}
}
