% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{functional_db_to_amplitude}
\alias{functional_db_to_amplitude}
\title{DB to Amplitude (functional)}
\usage{
functional_db_to_amplitude(x, ref = 1, power = 1)
}
\arguments{
\item{x}{(Tensor): Input tensor before being converted to power/amplitude scale.}

\item{ref}{(float): Reference which the output will be scaled by. (Default: \code{1.0})}

\item{power}{(float): If power equals 1, will compute DB to power. If 0.5, will compute
DB to amplitude. (Default: \code{1.0})}
}
\value{
\code{tensor}: Output tensor in power/amplitude scale.
}
\description{
Turn a tensor from the decibel scale to the power/amplitude scale.
}
