% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{functional_deemph_biquad}
\alias{functional_deemph_biquad}
\title{ISO 908 CD De-emphasis IIR Filter (functional)}
\usage{
functional_deemph_biquad(waveform, sample_rate)
}
\arguments{
\item{waveform}{(Tensor): audio waveform of dimension of \verb{(..., time)}}

\item{sample_rate}{(int): sampling rate of the waveform, Allowed sample rate \code{44100} or \code{48000}}
}
\value{
Tensor: Waveform of dimension of \verb{(..., time)}
}
\description{
Apply ISO 908 CD de-emphasis (shelving) IIR filter.  Similar to SoX implementation.
}
\references{
\itemize{
\item \url{https://sox.sourceforge.net/sox.html}
\item \url{https://webaudio.github.io/Audio-EQ-Cookbook/audio-eq-cookbook.html}
}
}
