% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{functional_find_max_per_frame}
\alias{functional_find_max_per_frame}
\title{Find Max Per Frame (functional)}
\usage{
functional_find_max_per_frame(nccf, sample_rate, freq_high)
}
\description{
For each frame, take the highest value of NCCF,
apply centered median smoothing, and convert to frequency.
}
\details{
Note: If the max among all the lags is very close
to the first half of lags, then the latter is taken.
}
