% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{functional_gain}
\alias{functional_gain}
\title{Gain (functional)}
\usage{
functional_gain(waveform, gain_db = 1)
}
\arguments{
\item{waveform}{(Tensor): Tensor of audio of dimension (..., time).}

\item{gain_db}{(float, optional) Gain adjustment in decibels (dB) (Default: \code{1.0}).}
}
\value{
\code{tensor}: the whole waveform amplified by gain_db.
}
\description{
Apply amplification or attenuation to the whole waveform.
}
