% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{functional_magphase}
\alias{functional_magphase}
\title{Magnitude and Phase}
\usage{
functional_magphase(complex_tensor, power = 1)
}
\arguments{
\item{complex_tensor}{(Tensor): Tensor shape of \verb{(.., complex=2)}}

\item{power}{(float): Power of the norm. (Default: \code{1.0})}
}
\value{
list(\code{tensor}, \code{tensor}): The magnitude and phase of the complex tensor
}
\description{
Separate a complex-valued spectrogram with shape \verb{(.., 2)} into its magnitude and phase.
}
