% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{functional_mask_along_axis}
\alias{functional_mask_along_axis}
\title{Mask Along Axis}
\usage{
functional_mask_along_axis(specgram, mask_param, mask_value, axis)
}
\arguments{
\item{specgram}{(Tensor): Real spectrogram (channel, freq, time)}

\item{mask_param}{(int): Number of columns to be masked will be uniformly sampled from \link{0, mask_param}}

\item{mask_value}{(float): Value to assign to the masked columns}

\item{axis}{(int): Axis to apply masking on (2 -> frequency, 3 -> time)}
}
\value{
Tensor: Masked spectrogram of dimensions (channel, freq, time)
}
\description{
Apply a mask along \code{axis}. Mask will be applied from indices \verb{[v_0, v_0 + v)}, where
\code{v} is sampled from \code{uniform (0, mask_param)}, and \code{v_0} from \code{uniform(0, max_v - v)}.
All examples will have the same mask interval.
}
