% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{functional_median_smoothing}
\alias{functional_median_smoothing}
\title{Median Smoothing (functional)}
\usage{
functional_median_smoothing(indices, win_length)
}
\arguments{
\item{indices}{(Tensor)}

\item{win_length}{(int)}
}
\description{
Apply median smoothing to the 1D tensor over the given window.
}
