% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{functional_mu_law_encoding}
\alias{functional_mu_law_encoding}
\title{Mu Law Encoding}
\usage{
functional_mu_law_encoding(x, quantization_channels)
}
\arguments{
\item{x}{(Tensor): Input tensor}

\item{quantization_channels}{(int): Number of channels}
}
\value{
\code{tensor}: Input after mu-law encoding
}
\description{
Encode signal based on mu-law companding.  For more info see
the \href{https://en.wikipedia.org/wiki/M-law_algorithm}{Wikipedia Entry}
}
\details{
This algorithm assumes the signal has been scaled to between -1 and 1 and
returns a signal encoded with values from 0 to quantization_channels - 1.
}
