% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{functional_overdrive}
\alias{functional_overdrive}
\title{Overdrive Effect (functional)}
\usage{
functional_overdrive(waveform, gain = 20, colour = 20)
}
\arguments{
\item{waveform}{(Tensor): audio waveform of dimension of \verb{(..., time)}}

\item{gain}{(float): desired gain at the boost (or attenuation) in dB
Allowed range of values are 0 to 100}

\item{colour}{(float):  controls the amount of even harmonic content in
the over-driven output. Allowed range of values are 0 to 100}
}
\value{
Tensor: Waveform of dimension of \verb{(..., time)}
}
\description{
Apply a overdrive effect to the audio. Similar to SoX implementation.
This effect applies a non linear distortion to the audio signal.
}
\references{
\itemize{
\item \url{http://sox.sourceforge.net/sox.html}
}
}
