% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{functional_riaa_biquad}
\alias{functional_riaa_biquad}
\title{RIAA Vinyl Playback Equalisation (functional)}
\usage{
functional_riaa_biquad(waveform, sample_rate)
}
\arguments{
\item{waveform}{(Tensor): audio waveform of dimension of \verb{(..., time)}}

\item{sample_rate}{(int): sampling rate of the waveform, e.g. 44100 (Hz).
Allowed sample rates in Hz : \code{44100},\code{48000},\code{88200},\code{96000}}
}
\value{
\code{tensor}: Waveform of dimension of \verb{(..., time)}
}
\description{
Apply RIAA vinyl playback equalisation.  Similar to SoX implementation.
}
\references{
\itemize{
\item \url{https://sox.sourceforge.net/sox.html}
\item \url{https://webaudio.github.io/Audio-EQ-Cookbook/audio-eq-cookbook.html}
}
}
