% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{functional_sliding_window_cmn}
\alias{functional_sliding_window_cmn}
\title{sliding-window Cepstral Mean Normalization}
\usage{
functional_sliding_window_cmn(
  waveform,
  cmn_window = 600,
  min_cmn_window = 100,
  center = TRUE,
  norm_vars = TRUE
)
}
\arguments{
\item{waveform}{(Tensor): Tensor of audio of dimension (..., freq, time)}

\item{cmn_window}{(int, optional): Window in frames for running average CMN computation (int, default = 600)}

\item{min_cmn_window}{(int, optional):  Minimum CMN window used at start of decoding (adds latency only at start).
Only applicable if center == \code{FALSE}, ignored if center==\code{TRUE}  (int, default = 100)}

\item{center}{(bool, optional): If \code{TRUE}, use a window centered on the current frame
(to the extent possible, modulo end effects). If \code{FALSE}, window is to the left. (bool, default = \code{FALSE})}

\item{norm_vars}{(bool, optional): If \code{TRUE}, normalize variance to one. (bool, default = \code{FALSE})}
}
\value{
\code{tensor}: Tensor of freq of dimension (..., frame)
}
\description{
Apply sliding-window cepstral mean  (and optionally variance) normalization per utterance.
}
