% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-common.R
\name{info}
\alias{info}
\title{Audio Information}
\usage{
info(filepath)
}
\arguments{
\item{filepath}{(str) path to the mp3/wav file.}
}
\value{
list(sample_rate, channels, samples)
}
\description{
Retrive metadata from mp3 or wave file without load the audio samples in memory.
}
