% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaldi.R
\name{kaldi__get_num_lr_output_samples}
\alias{kaldi__get_num_lr_output_samples}
\title{Linear Resample Output Samples}
\usage{
kaldi__get_num_lr_output_samples(input_num_samp, samp_rate_in, samp_rate_out)
}
\arguments{
\item{input_num_samp}{(int): The number of samples in the input}

\item{samp_rate_in}{(float): The original frequency of the signal}

\item{samp_rate_out}{(float): The desired frequency}
}
\value{
int: The number of output samples
}
\description{
Based on LinearResample::GetNumOutputSamples. LinearResample  (LR) means that
the output signal is at linearly spaced intervals  (i.e the output signal has a
frequency of \code{new_freq}). It uses sinc/bandlimited interpolation to upsample/downsample
the signal.
}
