% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-common.R
\name{list_audio_backends}
\alias{list_audio_backends}
\title{List available audio backends}
\usage{
list_audio_backends()
}
\value{
character vector with the list of available backends.
}
\description{
List available audio backends
}
