% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models-convtasnet.R
\name{model_conv_block}
\alias{model_conv_block}
\title{1D Convolutional block.}
\usage{
model_conv_block(
  io_channels,
  hidden_channels,
  kernel_size,
  padding,
  dilation = 1,
  no_residual = FALSE
)
}
\arguments{
\item{io_channels}{(int): The number of input/output channels, (B, Sc)}

\item{hidden_channels}{(int): The number of channels in the internal layers, H.}

\item{kernel_size}{(int): The convolution kernel size of the middle layer, P.}

\item{padding}{(int): Padding value of the convolution in the middle layer.}

\item{dilation}{(int): Dilation value of the convolution in the middle layer.}

\item{no_redisual}{(bool): Disable residual block/output.}
}
\description{
1D Convolutional block.
}
\details{
This implementation corresponds to the "non-causal" setting in the paper.
}
\section{Reference}{

\itemize{
\item Conv-TasNet: Surpassing Ideal Time--Frequency Magnitude Masking for Speech Separation
Luo, Yi and Mesgarani, Nima - \url{https://arxiv.org/abs/1809.07454}
}
}

