% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models-convtasnet.R
\name{model_conv_tasnet}
\alias{model_conv_tasnet}
\title{Conv-TasNet}
\usage{
model_conv_tasnet(
  num_sources = 2,
  enc_kernel_size = 16,
  enc_num_feats = 512,
  msk_kernel_size = 3,
  msk_num_feats = 128,
  msk_num_hidden_feats = 512,
  msk_num_layers = 8,
  msk_num_stacks = 3
)
}
\arguments{
\item{num_sources}{(int): The number of sources to split.}

\item{enc_kernel_size}{(int): The convolution kernel size of the encoder/decoder, L.}

\item{enc_num_feats}{(int): The feature dimensions passed to mask generator, N.}

\item{msk_kernel_size}{(int): The convolution kernel size of the mask generator, P.}

\item{msk_num_feats}{(int): The input/output feature dimension of conv block in the mask generator, (B, Sc).}

\item{msk_num_hidden_feats}{(int): The internal feature dimension of conv block of the mask generator, H.}

\item{msk_num_layers}{(int): The number of layers in one conv block of the mask generator, X.}

\item{msk_num_stacks}{(int): The numbr of conv blocks of the mask generator, R.}
}
\value{
Tensor: 3D Tensor with shape \link{batch, channel==num_sources, frames}
}
\description{
A fully-convolutional time-domain audio separation network. Perform source separation.
Generate audio source waveforms. This implementation corresponds to the "non-causal"
setting in the paper.
}
\details{
forward param:
\itemize{
\item input  (Tensor): 3D Tensor with shape \link{batch, channel==1, frames}
\item .align_num_frames_with_strides
}
}
\section{Reference}{

\itemize{
\item Conv-TasNet: Surpassing Ideal Time--Frequency Magnitude Masking for Speech Separation
Luo, Yi and Mesgarani, Nima \url{https://arxiv.org/abs/1809.07454}
}
}

