% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models-convtasnet.R
\name{model_mask_generator}
\alias{model_mask_generator}
\title{Temporal Convolution Network (TCN) Separation Module}
\usage{
model_mask_generator(
  input_dim,
  num_sources,
  kernel_size,
  num_feats,
  num_hidden,
  num_layers,
  num_stacks
)
}
\arguments{
\item{input_dim}{(int): Input feature dimension, N.}

\item{num_sources}{(int): The number of sources to separate.}

\item{kernel_size}{(int): The convolution kernel size of conv blocks, P.}

\item{num_hidden}{(int): Intermediate feature dimention of conv blocks, H}

\item{num_layers}{(int): The number of conv blocks in one stack, X.}

\item{num_stacks}{(int): The number of conv block stacks, R.}

\item{num_featrs}{(int): Input/output feature dimenstion of conv blocks, (B, Sc).}
}
\value{
(Tensor): shape \link{batch, num_sources, features, frames}
}
\description{
Generates masks for separation. This implementation corresponds to
the "non-causal" setting in the paper.
}
\details{
forward param
\itemize{
\item input (Tensor): 3D Tensor with shape \link{batch, features, frames}
}
}
\section{References}{

- Conv-TasNet: Surpassing Ideal Time--Frequency Magnitude Masking for Speech Separation
Luo, Yi and Mesgarani, Nima \url{https://arxiv.org/abs/1809.07454}
}

