% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-common.R
\name{mp3_info}
\alias{mp3_info}
\title{MP3 Information}
\usage{
mp3_info(filepath)
}
\arguments{
\item{filepath}{(chr) path to mp3 file}
}
\value{
AudioMetaData: sample_rate, channels, samples
}
\description{
Retrive metadata from mp3 without load the audio samples in memory.
}
\examples{
mp3_path <- system.file("sample_audio_1.mp3", package = "torchaudio")
mp3_info(mp3_path)

}
