% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-common.R
\name{set_audio_backend}
\alias{set_audio_backend}
\title{Set the backend for I/O operation}
\usage{
set_audio_backend(backend)
}
\arguments{
\item{backend}{(str or function): one of \code{av_loader},
\code{audiofile_loader} or \code{tuneR_loader}.}
}
\value{
invisible(NULL).

It will set \code{torchaudio.loader} option:``
options(torchaudio.loader = rlang::as_function(backend))
}
\description{
Set the backend for I/O operation
}
