% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-utils.R
\name{strip}
\alias{strip}
\title{Strip}
\usage{
strip(str)
}
\description{
removes any leading (spaces at the beginning) and trailing (spaces at the end) characters.
Analog to strip() string method from Python.
}
\keyword{internal}
