% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-common.R
\name{torchaudio_info}
\alias{torchaudio_info}
\title{Audio Information}
\usage{
torchaudio_info(filepath)
}
\arguments{
\item{filepath}{(str) path to the audio file.}
}
\value{
AudioMetaData: an R6 class with fields sample_rate, channels, samples.
}
\description{
Retrieve audio metadata.
}
\examples{
path <- system.file("waves_yesno/1_1_0_1_1_0_1_1.wav", package = "torchaudio")
torchaudio_info(path)

}
