% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{transform__axismasking}
\alias{transform__axismasking}
\title{Axis Masking}
\usage{
transform__axismasking(mask_param, axis, iid_masks)
}
\arguments{
\item{mask_param}{(int): Maximum possible length of the mask.}

\item{axis}{(int): What dimension the mask is applied on.}

\item{iid_masks}{(bool): Applies iid masks to each of the examples in the batch dimension.
This option is applicable only when the input tensor is 4D.}
}
\value{
Tensor: Masked spectrogram of dimensions (..., freq, time).
}
\description{
Apply masking to a spectrogram.
}
\details{
forward param:
specgram  (Tensor): Tensor of dimension (..., freq, time).

mask_value  (float): Value to assign to the masked columns.
}
