% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tranforms.R
\name{transform_amplitude_to_db}
\alias{transform_amplitude_to_db}
\title{Amplitude to DB}
\usage{
transform_amplitude_to_db(stype = "power", top_db = NULL)
}
\arguments{
\item{stype}{(str, optional): scale of input tensor ('power' or 'magnitude'). The
power being the elementwise square of the magnitude. (Default: \code{'power'})}

\item{top_db}{(float or NULL, optional): Minimum negative cut-off in decibels. A reasonable number
is 80. (Default: \code{NULL})}
}
\value{
\code{tensor}: Output tensor in decibel scale
}
\description{
Turn a tensor from the power/amplitude scale to the decibel scale.
}
\details{
This output depends on the maximum value in the input tensor, and so
may return different values for an audio clip split into snippets vs. a
a full clip.
}
