% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tranforms.R
\name{transform_complex_norm}
\alias{transform_complex_norm}
\title{Complex Norm}
\usage{
transform_complex_norm(power = 1)
}
\arguments{
\item{power}{(float, optional): Power of the norm. (Default: to \code{1.0})}
}
\value{
Tensor: norm of the input tensor, shape of \verb{(..., )}.
}
\description{
Compute the norm of complex tensor input.
}
\details{
forward param:
complex_tensor  (Tensor): Tensor shape of \verb{(..., complex=2)}.
}
