% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tranforms.R
\name{transform_fade}
\alias{transform_fade}
\title{Add a fade in and/or fade out to an waveform.}
\usage{
transform_fade(fade_in_len = 0, fade_out_len = 0, fade_shape = "linear")
}
\arguments{
\item{fade_in_len}{(int, optional): Length of fade-in (time frames). (Default: \code{0})}

\item{fade_out_len}{(int, optional): Length of fade-out (time frames). (Default: \code{0})}

\item{fade_shape}{(str, optional): Shape of fade. Must be one of: "quarter_sine",
"half_sine", "linear", "logarithmic", "exponential".  (Default: \code{"linear"})}

\item{waveform}{(Tensor): Tensor of audio of dimension (..., time).}
}
\value{
Tensor: Tensor of audio of dimension (..., time).
}
\description{
Add a fade in and/or fade out to an waveform.
}
