% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tranforms.R
\name{transform_frequencymasking}
\alias{transform_frequencymasking}
\title{Frequency-domain Masking}
\usage{
transform_frequencymasking(freq_mask_param, iid_masks)
}
\arguments{
\item{freq_mask_param}{(int): maximum possible length of the mask.
Indices uniformly sampled from [0, freq_mask_param).}

\item{iid_masks}{(bool, optional): whether to apply different masks to each
example/channel in the batch.  (Default: \code{FALSE})
This option is applicable only when the input tensor is 4D.}
}
\description{
Apply masking to a spectrogram in the frequency domain.
}
