% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tranforms.R
\name{transform_resample}
\alias{transform_resample}
\title{Signal Resample}
\usage{
transform_resample(
  orig_freq = 16000,
  new_freq = 16000,
  resampling_method = "sinc_interpolation"
)
}
\arguments{
\item{orig_freq}{(float, optional): The original frequency of the signal. (Default: \code{16000})}

\item{new_freq}{(float, optional): The desired frequency. (Default: \code{16000})}

\item{resampling_method}{(str, optional): The resampling method. (Default: \code{'sinc_interpolation'})}
}
\value{
Tensor: Output signal of dimension (..., time).
}
\description{
Resample a signal from one frequency to another. A resampling method can be given.
}
\details{
forward param:
waveform  (Tensor): Tensor of audio of dimension (..., time).
}
