% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{transform_timemasking}
\alias{transform_timemasking}
\title{Time-domain Masking}
\usage{
transform_timemasking(time_mask_param, iid_masks)
}
\arguments{
\item{time_mask_param}{(int): maximum possible length of the mask.
Indices uniformly sampled from [0, time_mask_param).}

\item{iid_masks}{(bool, optional): whether to apply different masks to each
example/channel in the batch.  (Default: \code{FALSE})
This option is applicable only when the input tensor is 4D.}
}
\value{
not implemented yet.
}
\description{
Apply masking to a spectrogram in the time domain.
}
