% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{transform_vol}
\alias{transform_vol}
\title{Add a volume to an waveform.}
\usage{
transform_vol(gain, gain_type = "amplitude")
}
\arguments{
\item{gain}{(float): Interpreted according to the given gain_type:
If \code{gain_type} = \code{amplitude}, \code{gain} is a positive amplitude ratio.
If \code{gain_type} = \code{power}, \code{gain} is a power  (voltage squared).
If \code{gain_type} = \code{db}, \code{gain} is in decibels.}

\item{gain_type}{(str, optional): Type of gain. One of: \code{amplitude}, \code{power}, \code{db} (Default: \code{amplitude})}
}
\value{
Tensor: Tensor of audio of dimension (..., time).
}
\description{
Add a volume to an waveform.
}
\details{
forward param:
waveform  (Tensor): Tensor of audio of dimension (..., time).
}
