% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-tuneR.R
\name{tuneR_loader}
\alias{tuneR_loader}
\title{tuneR_loader}
\usage{
tuneR_loader(filepath, offset = 0L, duration = 0L, unit = c("samples", "time"))
}
\arguments{
\item{filepath}{(str) path to the audio file.}

\item{offset}{(num) the sample (or the second if unit = 'time') where the audio should start.}

\item{duration}{(num) how many samples (or how many seconds if unit = 'time') should be extracted.}

\item{unit}{(str) 'samples' or 'time'}
}
\description{
Load an audio located at 'filepath' using tuneR package.
}
