% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-utils.R
\name{walk_files}
\alias{walk_files}
\title{List recursively all files ending with a suffix at a given root}
\usage{
walk_files(root, suffix, prefix = FALSE, remove_suffix = FALSE)
}
\arguments{
\item{root}{(str): Path to directory whose folders need to be listed}

\item{suffix}{(str or tuple): Suffix of the files to match, e.g. '.png' or ('.jpg', '.png').
It uses the Python "str.endswith" method and is passed directly}

\item{prefix}{(bool, optional): If TRUE, prepends the full path to each result, otherwise
only returns the name of the files found  (Default: \code{FALSE})}

\item{remove_suffix}{(bool, optional): If TRUE, removes the suffix to each result defined in suffix,
otherwise will return the result as found  (Default: \code{FALSE}).}
}
\description{
List recursively all files ending with a suffix at a given root
}
\keyword{internal}
