% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-common.R
\name{wav_info}
\alias{wav_info}
\title{Wave Information}
\usage{
wav_info(filepath)
}
\arguments{
\item{filepath}{(chr) path to wav file}
}
\value{
AudioMetaData: sample_rate, channels, samples
}
\description{
Retrive metadata from wav without load the audio samples in memory.
}
\examples{
wav_path <- system.file("waves_yesno/1_1_0_1_1_0_1_1.wav", package = "torchaudio")
wav_info(wav_path)

}
