% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torchaudio.R
\name{wave_to_tensor}
\alias{wave_to_tensor}
\title{Load Audio File}
\usage{
wave_to_tensor(
  wave_obj,
  out = out,
  normalization = normalization,
  channels_first = channels_first
)
}
\arguments{
\item{wave_obj}{(Wave): Wave object, usually an output from \link[tuneR:readMP3]{tuneR::readMP3}, \link[tuneR:readWave]{tuneR::readWave} or \link[monitoR:readMP3]{monitoR::readMP3}.}

\item{out}{(Tensor): An optional output tensor to use instead of creating one. (Default: \code{NULL})}

\item{normalization}{(bool, float or function): Optional normalization.
If boolean \code{TRUE}, then output is divided by \code{2^31}.
Assuming the input is signed 32-bit audio, this normalizes to \verb{[-1, 1]}.
If \code{numeric}, then output is divided by that number.
If \code{function}, then the output is passed as a paramete to the given function,
then the output is divided by the result. (Default: \code{TRUE})}

\item{channels_first}{(bool): Set channels first or length first in result. (Default: \code{TRUE})}
}
\value{
\preformatted{list(Tensor, int): An output tensor of size `[C x L]` or `[L x C]` where
    L is the number of audio frames and
    C is the number of channels.
    An integer which is the sample rate of the audio (as listed in the metadata of the file)
}
}
\description{
Loads an audio file from disk into a tensor
}
