% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-yesno.R
\name{yesno_dataset}
\alias{yesno_dataset}
\title{YesNo Dataset}
\usage{
yesno_dataset(
  root,
  url = "http://www.openslr.org/resources/1/waves_yesno.tar.gz",
  folder_in_archive = "waves_yesno",
  download = FALSE,
  transform = NULL,
  target_transform = NULL
)
}
\arguments{
\item{root}{(str): Path to the directory where the dataset is found or downloaded.}

\item{url}{(str, optional): The URL to download the dataset from.
(default: \code{"[http://www.openslr.org/resources/1/waves_yesno.tar.gz]()"})}

\item{folder_in_archive}{(str, optional): The top-level directory of the dataset.  (default: \code{"waves_yesno"})}

\item{download}{(bool, optional): Whether to download the dataset if it is not found at root path.  (default: \code{FALSE}).}

\item{transform}{(callable, optional): Optional transform applied on waveform. (default: \code{NULL})}

\item{target_transform}{(callable, optional): Optional transform applied on utterance. (default: \code{NULL})}
}
\value{
tuple: \verb{(waveform, sample_rate, labels)}
}
\description{
Create a Dataset for YesNo
}
