% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-common.R
\name{torchaudio_load}
\alias{torchaudio_load}
\title{Load Audio File}
\usage{
torchaudio_load(
  filepath,
  offset = 0L,
  duration = -1L,
  unit = c("samples", "time")
)
}
\arguments{
\item{filepath}{(str): Path to audio file}

\item{offset}{(int): Number of frames (or seconds) from the start of the file to begin data loading. (Default: \code{0})}

\item{duration}{(int): Number of frames (or seconds) to load.  \code{-1} to load everything after the offset. (Default: \code{-1})}

\item{unit}{(str): "sample" or "time". If "sample" duration and offset will be interpreted as frames, and as seconds otherwise.}
}
\description{
Loads an audio file from disk using the default loader (getOption("torchaudio.loader")).
}
